+++ site/content/examples/04-events/05-dom-event-forwarding/CustomButton.svelte
@@ -1,3 +1,7 @@
+<button on:click>
+	Click me
+</button>
+
 <style>
 	button {
 		height: 4rem;
@@ -16,7 +20,3 @@
 		color: #aaa;
 	}
 </style>
-
-<button on:click>
-	Click me
-</button>