+++ site/content/examples/12-svg/02-bar-chart/App.svelte
@@ -33,6 +33,42 @@
 	$: barWidth = innerWidth / xTicks.length;
 </script>
 
+<h2>US birthrate by year</h2>
+
+<div class="chart" bind:clientWidth={width} bind:clientHeight={height}>
+	<svg>
+		<!-- y axis -->
+		<g class="axis y-axis" transform="translate(0,{padding.top})">
+			{#each yTicks as tick}
+				<g class="tick tick-{tick}" transform="translate(0, {yScale(tick) - padding.bottom})">
+					<line x2="100%"></line>
+					<text y="-4">{tick} {tick === 20 ? ' per 1,000 population' : ''}</text>
+				</g>
+			{/each}
+		</g>
+
+		<!-- x axis -->
+		<g class="axis x-axis">
+			{#each points as point, i}
+				<g class="tick" transform="translate({xScale(i)},{height})">
+					<text x="{barWidth/2}" y="-4">{width > 380 ? point.year : formatMobile(point.year)}</text>
+				</g>
+			{/each}
+		</g>
+
+		<g class='bars'>
+			{#each points as point, i}
+				<rect
+					x="{xScale(i) + 2}"
+					y="{yScale(point.birthrate)}"
+					width="{barWidth - 4}"
+					height="{height - padding.bottom - yScale(point.birthrate)}"
+				></rect>
+			{/each}
+		</g>
+	</svg>
+</div>
+
 <style>
 	h2 {
 		text-align: center;
@@ -80,39 +116,3 @@
 		opacity: 0.65;
 	}
 </style>
-
-<h2>US birthrate by year</h2>
-
-<div class="chart" bind:clientWidth={width} bind:clientHeight={height}>
-	<svg>
-		<!-- y axis -->
-		<g class="axis y-axis" transform="translate(0,{padding.top})">
-			{#each yTicks as tick}
-				<g class="tick tick-{tick}" transform="translate(0, {yScale(tick) - padding.bottom})">
-					<line x2="100%"></line>
-					<text y="-4">{tick} {tick === 20 ? ' per 1,000 population' : ''}</text>
-				</g>
-			{/each}
-		</g>
-
-		<!-- x axis -->
-		<g class="axis x-axis">
-			{#each points as point, i}
-				<g class="tick" transform="translate({xScale(i)},{height})">
-					<text x="{barWidth/2}" y="-4">{width > 380 ? point.year : formatMobile(point.year)}</text>
-				</g>
-			{/each}
-		</g>
-
-		<g class='bars'>
-			{#each points as point, i}
-				<rect
-					x="{xScale(i) + 2}"
-					y="{yScale(point.birthrate)}"
-					width="{barWidth - 4}"
-					height="{height - padding.bottom - yScale(point.birthrate)}"
-				></rect>
-			{/each}
-		</g>
-	</svg>
-</div>