+++ site/content/tutorial/16-special-elements/02-svelte-component/app-b/RedThing.svelte
@@ -1,5 +1,7 @@
-<style>
-	strong { color: red; }
-</style>
+<strong>Red thing</strong>
 
-<strong>red thing</strong>
+<style>
+	strong {
+		color: red;
+	}
+</style>