+++ src/compiler/compile/index.ts
@@ -15,6 +15,7 @@ const valid_options = [
 	'sourcemap',
 	'generate',
 	'errorMode',
+	'varsReport',
 	'outputFilename',
 	'cssOutputFilename',
 	'sveltePath',