+++ src/compiler/compile/utils/scope.ts
@@ -1,8 +1,8 @@
 import { Node } from 'estree';
-import { analyze, Scope, extract_names, extract_identifiers } from 'periscopic';
+import { Node as PeriscopicNode, analyze, Scope, extract_names, extract_identifiers } from 'periscopic';
 
 export function create_scopes(expression: Node) {
-	return analyze(expression);
+	return analyze(expression as PeriscopicNode);
 }
 
 export { Scope, extract_names, extract_identifiers };