+++ test/runtime/samples/component-namespaced/_config.js
@@ -1,5 +1,3 @@
-import * as path from 'path';
-
 export default {
 	props: {
 		a: 1
@@ -8,10 +6,6 @@ export default {
 	html: `
 		<p>foo 1</p>
 	`,
-
-	before_test() {
-		delete require.cache[path.resolve(__dirname, 'components.js')];
-	},
 
 	test({ assert, component, target }) {
 		component.a = 2;